C++ 17 compatible

+++ application/components/LCR/LCR_Hub.h
@@ -16,11 +16,11 @@ public:
     inline LCR_Hub()
     {}
 
-    void toFld(UINT32 id, INT32 state) const;
-    void toGui(UINT32 id, INT32 state, INT32 ubk) const;
+    void toFld(size_t id, INT32 state) const;
+    void toGui(size_t id, INT32 state, INT32 ubk) const;
 
-    void fromDsp(UINT32 pos, const FldState& tele) const;
-    void fromDsp(UINT32 pos, const GuiCmd&   tele) const;
+    void fromDsp(size_t pos, const FldState& tele) const;
+    void fromDsp(size_t pos, const GuiCmd&   tele) const;
 
     IL_INSTANCE_DEC(LCR_Hub)
