+++ application/components/LCR/src/LCR_Hub.cpp
@@ -5,7 +5,7 @@
 
 INSTANCE_DEF(LCR_Hub)
 
-void LCR_Hub::fromDsp(const UINT32 pos, const FldState& tele) const
+void LCR_Hub::fromDsp(const size_t pos, const FldState& tele) const
 {
     I_LCR_Provider& prov = IL::getLCR_Provider();
     if (prov.has(pos))
@@ -16,7 +16,7 @@ void LCR_Hub::fromDsp(const UINT32 pos, const FldState& tele) const
     {pass();}
 }
 
-void LCR_Hub::fromDsp(const UINT32 pos, const GuiCmd&   tele) const
+void LCR_Hub::fromDsp(const size_t pos, const GuiCmd&   tele) const
 {
     I_LCR_Provider& prov = IL::getLCR_Provider();
     if (prov.has(pos))
@@ -27,7 +27,7 @@ void LCR_Hub::fromDsp(const UINT32 pos, const GuiCmd&   tele) const
     {pass();}
 }
 
-void LCR_Hub::toFld(const UINT32 id, const INT32 state) const
+void LCR_Hub::toFld(const size_t id, const INT32 state) const
 {
     static CmdFld tele;
     Mem::zero(tele);
@@ -35,7 +35,7 @@ void LCR_Hub::toFld(const UINT32 id, const INT32 state) const
     IL::getDispatcher().dispatch(id, tele);
 }
 
-void LCR_Hub::toGui(const UINT32 id, const INT32 state, const INT32 ubk) const
+void LCR_Hub::toGui(const size_t id, const INT32 state, const INT32 ubk) const
 {
     static StateGui tele;
     Mem::zero(tele);
